+++ gdsfactory/components/compass.py
@@ -10,7 +10,7 @@ from gdsfactory.types import Layer
 def compass(
     size: Tuple[float, float] = (4.0, 2.0),
     layer: Layer = gf.LAYER.WG,
-    prefix: str = "e",
+    port_type: str = "electrical",
 ) -> Component:
     """Rectangular contact pad with centered ports on rectangle edges
     (north, south, east, and west)
@@ -20,6 +20,7 @@ def compass(
         layer: tuple (int, int)
 
     """
+    prefix = "e"
 
     c = gf.Component()
     r = gf.components.rectangle(size=size, layer=layer)
@@ -34,12 +35,23 @@ def compass(
         width=dy,
         orientation=180,
         layer=layer,
+        port_type=port_type,
     )
     c.add_port(
-        name=f"{prefix}2", midpoint=[0, dy / 2], width=dx, orientation=90, layer=layer
+        name=f"{prefix}2",
+        midpoint=[0, dy / 2],
+        width=dx,
+        orientation=90,
+        layer=layer,
+        port_type=port_type,
     )
     c.add_port(
-        name=f"{prefix}3", midpoint=[dx / 2, 0], width=dy, orientation=0, layer=layer
+        name=f"{prefix}3",
+        midpoint=[dx / 2, 0],
+        width=dy,
+        orientation=0,
+        layer=layer,
+        port_type=port_type,
     )
     c.add_port(
         name=f"{prefix}4",
@@ -47,10 +59,12 @@ def compass(
         width=dx,
         orientation=-90,
         layer=layer,
+        port_type=port_type,
     )
+    c.auto_rename_ports()
     return c
 
 
 if __name__ == "__main__":
-    c = compass(size=(1, 2), layer=gf.LAYER.WG)
+    c = compass(size=(1, 2), layer=gf.LAYER.WG, port_type="optical")
     c.show()